perf(verification): one-point rule for `x == c` premises in VC flattening
When `flatten` introduces a fresh binder `x` whose premise is exactly
`x == c`, apply the one-point rule: substitute `c` for `x` in the body and
drop the binder + premise entirely instead of handing Z3 one more
universally quantified variable and assumption.

`forall x:b, (x == c) => P` is logically equivalent to `P[x := c]` since the
freshly introduced `x` never occurs free in `c`, so this is always sound.
This runs in the live verification path (`smt_valid` -> `flatten`); the
pre-existing equality elimination in `simplify_constraint` only ran in the
error-reporting path, so the actual SMT query never benefited from it.

`substitution_in_constraint` is moved from `helpers.py` to `vcs.py` (next to
the `Constraint` datatypes) so both `smt.py` and `helpers.py` can import it
without an import cycle.

Speedup (isolated SMT verification path, one-point ON vs OFF):
  super-mario (6 chunks)  1.11x   (244 `==` premises, 114 eliminated)
  super-mario (8 chunks)  1.06x
  AVL depth 4/5           1.00x   (only 5% of binders carry an `==` premise;
                                   cost is dominated by if-then-else height
                                   arithmetic the rule cannot touch)

The win is modest: Z3 already eliminates trivial equalities during
preprocessing, and SMT solving is ~18% of total typecheck time. The value
here is a sound, always-smaller VC rather than a large wall-clock gain.
All 1263 tests pass.
